changed sizes to size_t type

+++ LibSource/ComplexFloatArray.cpp
@@ -18,7 +18,7 @@ void ComplexFloatArray::getMagnitudeValues(FloatArray destination){
 #ifdef ARM_CORTEX
   arm_cmplx_mag_f32((float*)data, (float*)destination, size);
 #else
-  for(int i=0; i<size; i++){
+  for(size_t i=0; i<size; i++){
     destination[i]=mag(i);
   }
 #endif
@@ -43,7 +43,7 @@ void ComplexFloatArray::getMagnitudeSquaredValues(FloatArray destination){
 #ifdef ARM_CORTEX
   arm_cmplx_mag_squared_f32((float*)data, (float*)destination, size);
 #else
-  for(int i=0; i<size; i++){
+  for(size_t i=0; i<size; i++){
     destination[i]=mag2(i);
   }
 #endif  
@@ -52,7 +52,7 @@ void ComplexFloatArray::getMagnitudeSquaredValues(FloatArray destination){
 void ComplexFloatArray::getPolar(FloatArray magnitude, FloatArray phase){
   ASSERT(magnitude.getSize()>=size, "Wrong array size");
   ASSERT(phase.getSize()>=size, "Wrong array size");
-  for(int i=0; i<size; i++){
+  for(size_t i=0; i<size; i++){
     magnitude[i] = data[i].getMagnitude();
     phase[i] = data[i].getPhase();
   }
@@ -60,7 +60,7 @@ void ComplexFloatArray::getPolar(FloatArray magnitude, FloatArray phase){
 
 void ComplexFloatArray::getPhaseValues(FloatArray destination){
   ASSERT(destination.getSize()>=size, "Wrong array size");
-  for(int i=0; i<size; i++)
+  for(size_t i=0; i<size; i++)
     destination[i] = data[i].getPhase();
 }
 
@@ -73,7 +73,7 @@ void ComplexFloatArray::complexDotProduct(ComplexFloatArray operand2, ComplexFlo
   float *pSrcB=(float*)operand2;
   float realResult=0;    
   float imagResult=0;    
-  for(int n=0; n<size; n++) {    
+  for(size_t n=0; n<size; n++) {    
       realResult += pSrcA[(2*n)+0]*pSrcB[(2*n)+0] - pSrcA[(2*n)+1]*pSrcB[(2*n)+1];    
       imagResult += pSrcA[(2*n)+0]*pSrcB[(2*n)+1] + pSrcA[(2*n)+1]*pSrcB[(2*n)+0];    
   }
@@ -91,7 +91,7 @@ void ComplexFloatArray::complexByComplexMultiplication(ComplexFloatArray operand
   float *pSrcA=(float*)data;
   float *pSrcB=(float*)operand2;
   float *pDst=(float*)result;
-  for(int n=0; n<size; n++) {        
+  for(size_t n=0; n<size; n++) {        
     pDst[(2*n)+0] = pSrcA[(2*n)+0] * pSrcB[(2*n)+0] - pSrcA[(2*n)+1] * pSrcB[(2*n)+1];        
     pDst[(2*n)+1] = pSrcA[(2*n)+0] * pSrcB[(2*n)+1] + pSrcA[(2*n)+1] * pSrcB[(2*n)+0];        
   }        
@@ -103,7 +103,7 @@ void ComplexFloatArray::add(ComplexFloatArray operand2, ComplexFloatArray destin
 #ifdef ARM_CORTEX
   arm_add_f32((float*)data, (float*)operand2.data, (float*)destination.data, size*2);
 #else
-  for(int n=0; n<size; n++){
+  for(size_t n=0; n<size; n++){
     destination[n].re = data[n].re + operand2[n].re;
     destination[n].im = data[n].im + operand2[n].im;
   }
@@ -119,7 +119,7 @@ void ComplexFloatArray::subtract(ComplexFloatArray operand2, ComplexFloatArray d
 #ifdef ARM_CORTEX
   arm_sub_f32((float*)data, (float*)operand2.data, (float*)destination.data, size*2);
 #else
-  for(int n=0; n<size; n++){
+  for(size_t n=0; n<size; n++){
     destination[n].re = data[n].re - operand2[n].re;
     destination[n].im = data[n].im - operand2[n].im;
   }
@@ -138,7 +138,7 @@ void ComplexFloatArray::getComplexConjugateValues(ComplexFloatArray destination)
 #else
   float *pSrc=(float*)data;
   float *pDst=(float *)destination;
-  for(int n=0; n<size; n++) {        
+  for(size_t n=0; n<size; n++) {        
     pDst[(2*n)+0] = pSrc[(2*n)+0];     // real part        
     pDst[(2*n)+1] = -pSrc[(2*n)+1];    // imag part        
 }   
@@ -154,7 +154,7 @@ void ComplexFloatArray::complexByRealMultiplication(FloatArray operand2, Complex
   float *pSrcCmplx=(float*)data;
   float *pSrcReal=(float*)operand2;
   float *pCmplxDst=(float*)result;
-  for(int n=0; n<size; n++) {
+  for(size_t n=0; n<size; n++) {
       pCmplxDst[(2*n)+0] = pSrcCmplx[(2*n)+0] * pSrcReal[n];        
       pCmplxDst[(2*n)+1] = pSrcCmplx[(2*n)+1] * pSrcReal[n];        
   }        
@@ -164,7 +164,7 @@ void ComplexFloatArray::complexByRealMultiplication(FloatArray operand2, Complex
 int ComplexFloatArray::getMaxMagnitudeIndex(){ //this is probably slower than getMagnitudeSquaredValues() and getMaxIndex() on it
   float maxMag=-1;
   int maxInd=-1;
-  for(int n=0; n<size; n++){
+  for(size_t n=0; n<size; n++){
     float magnitude=mag2(n); //uses mag2 which is faster
     if(magnitude>maxMag){
       maxMag=magnitude;
@@ -174,14 +174,14 @@ int ComplexFloatArray::getMaxMagnitudeIndex(){ //this is probably slower than ge
   return maxInd;
 }
 
-ComplexFloatArray ComplexFloatArray::subArray(int offset, int length){
+ComplexFloatArray ComplexFloatArray::subArray(int offset, size_t length){
   ASSERT(size >= offset+length, "Array too small");
   return ComplexFloatArray(data+offset, length);
 }
 
 float ComplexFloatArray::getMaxMagnitudeValue(){ //this is probably slower than getMagnitudeSquaredValues() and getMaxValue() on it
   float maxMag=-1;
-  for(int n=0; n<size; n++){
+  for(size_t n=0; n<size; n++){
     float mag=this->mag2(n);
     if(mag>maxMag){
       maxMag=mag;
@@ -192,13 +192,13 @@ float ComplexFloatArray::getMaxMagnitudeValue(){ //this is probably slower than 
 }
 
 void ComplexFloatArray::getRealValues(FloatArray buf){
-  for(int n=0; n<size; n++){
+  for(size_t n=0; n<size; n++){
     buf[n]=data[n].re;
   }
 }
 
 void ComplexFloatArray::getImaginaryValues(FloatArray buf){
-  for(int n=0; n<size; n++){
+  for(size_t n=0; n<size; n++){
     buf[n]=data[n].im;
   }
 }
@@ -208,14 +208,14 @@ void ComplexFloatArray::scale(float factor){
 #ifdef ARM_CORTEX  
   arm_scale_f32((float*)data, factor, (float*)data, size*2); //the *2 accounts for the fact that both real and imaginary parts are scaled
 #else
-  for(int n=0; n<size; n++){
+  for(size_t n=0; n<size; n++){
     data[n].re *= factor;
     data[n].im *= factor;
   }
 #endif
 }
 
-ComplexFloatArray ComplexFloatArray::create(int size){
+ComplexFloatArray ComplexFloatArray::create(size_t size){
   return ComplexFloatArray(new ComplexFloat[size], size);
 }
 
@@ -226,7 +226,7 @@ void ComplexFloatArray::destroy(ComplexFloatArray array){
 /* Copies real values from a FloatArray, sets imaginary values to 0
  */
 void ComplexFloatArray::copyFrom(FloatArray source){
-  for(int n=0; n<min(size,source.getSize()); n++){
+  for(size_t n=0; n<min(size,source.getSize()); n++){
     data[n].re=source[n];
     data[n].im=0;
   }
@@ -234,7 +234,7 @@ void ComplexFloatArray::copyFrom(FloatArray source){
 
 /* Copies real part to a FloatArray */
 void ComplexFloatArray::copyTo(FloatArray destination){
-  for(int n=0; n<min(size,destination.getSize()); n++){
+  for(size_t n=0; n<min(size,destination.getSize()); n++){
     destination[n]=data[n].re;
   }
 }
@@ -248,26 +248,26 @@ void ComplexFloatArray::copyFrom(ComplexFloatArray source){
   copyFrom(source, min(size, source.getSize()));
 }
 
-void ComplexFloatArray::copyTo(ComplexFloat* other, int length){
+void ComplexFloatArray::copyTo(ComplexFloat* other, size_t length){
   ASSERT(size >= length, "Array too small");
 /// @note When built for ARM Cortex-M processor series, this method uses the optimized <a href="http://www.keil.com/pack/doc/CMSIS/General/html/index.html">CMSIS library</a>
 #ifdef ARM_CORTEX
   arm_copy_f32((float *)data, (float *)other, length*sizeof(ComplexFloat)/sizeof(float));
 #else
-  for(int n=0; n<length; n++){
+  for(size_t n=0; n<length; n++){
     other[n].re=data[n].re;
     other[n].im=data[n].im;
   }
 #endif /* ARM_CORTEX */
 }
 
-void ComplexFloatArray::copyFrom(ComplexFloat* other, int length){
+void ComplexFloatArray::copyFrom(ComplexFloat* other, size_t length){
   ASSERT(size >= length, "Array too small");
 /// @note When built for ARM Cortex-M processor series, this method uses the optimized <a href="http://www.keil.com/pack/doc/CMSIS/General/html/index.html">CMSIS library</a>
 #ifdef ARM_CORTEX
   arm_copy_f32((float *)other, (float *)data, length*sizeof(ComplexFloat)/sizeof(float));  //note the *2 multiplier which accounts for real and imaginary parts
 #else
-  for(int n=0; n<length; n++){
+  for(size_t n=0; n<length; n++){
     data[n].re=other[n].re;
     data[n].im=other[n].im;
   }
@@ -288,7 +288,7 @@ void ComplexFloatArray::setAll(float value){
 /*END -- methods copied and adapted from ComplexFloatArray.cpp*/
 
 void ComplexFloatArray::setAll(ComplexFloat value){
-  for(int n=0; n<size; n++){
+  for(size_t n=0; n<size; n++){
     data[n].re=value.re;
     data[n].im=value.im;
   }
@@ -304,8 +304,8 @@ void ComplexFloatArray::setAll(float valueRe, float valueIm){
 void ComplexFloatArray::setPolar(FloatArray magnitude, FloatArray phase){
   setPolar(magnitude, phase, 0, size);
 }
-void ComplexFloatArray::setPolar(FloatArray magnitude, FloatArray phase, int offset, int count){
-  for(int n=offset; n<count+offset; n++){
+void ComplexFloatArray::setPolar(FloatArray magnitude, FloatArray phase, int offset, size_t count){
+  for(size_t n=offset; n<count+offset; n++){
     data[n].setPolar(magnitude[n], phase[n]);
   }
 }
@@ -313,35 +313,35 @@ void ComplexFloatArray::setPolar(FloatArray magnitude, FloatArray phase, int off
 void ComplexFloatArray::setPhase(FloatArray phase){
   setPhase(phase, 0, size);
 }
-void ComplexFloatArray::setPhase(FloatArray phase, int offset, int count){
-  for(int n=offset; n<count+offset; n++){
+void ComplexFloatArray::setPhase(FloatArray phase, int offset, size_t count){
+  for(size_t n=offset; n<count+offset; n++){
     data[n].setPhase(phase[n]);
   }
 }
 void ComplexFloatArray::setPhase(FloatArray phase, ComplexFloatArray destination){
   setPhase(phase, 0, size, destination);
 }
-void ComplexFloatArray::setPhase(FloatArray phase, int offset, int count, ComplexFloatArray destination){
+void ComplexFloatArray::setPhase(FloatArray phase, int offset, size_t count, ComplexFloatArray destination){
   ASSERT(destination.getSize()>=count+offset, "Wrong size");
-  for(int n=offset; n<count+offset; n++){
+  for(size_t n=offset; n<count+offset; n++){
     destination.getData()[n].setPolar(getData()[n].getMagnitude(), phase[n]);
   }
 }
 void ComplexFloatArray::setMagnitude(FloatArray magnitude){
   setMagnitude(magnitude, 0, size);
 }
-void ComplexFloatArray::setMagnitude(FloatArray magnitude, int offset, int count){
+void ComplexFloatArray::setMagnitude(FloatArray magnitude, int offset, size_t count){
   setMagnitude(magnitude, offset, count, *this);
 }
 void ComplexFloatArray::setMagnitude(FloatArray magnitude, ComplexFloatArray destination){
   setMagnitude(magnitude, 0, size, destination);
 }
-void ComplexFloatArray::setMagnitude(FloatArray magnitude, int offset, int count, ComplexFloatArray destination){
+void ComplexFloatArray::setMagnitude(FloatArray magnitude, int offset, size_t count, ComplexFloatArray destination){
   ASSERT(getSize()==magnitude.getSize(),"wrong size0");
   ASSERT(getSize()==destination.getSize(),"wrong size1");
   ASSERT(offset+count<=destination.getSize(), "Wrong size2");
   ASSERT(offset+count<=getSize(), "Wrong size3");
-  for(int n=offset; n<count+offset; n++){
+  for(size_t n=offset; n<count+offset; n++){
     destination.getData()[n].setPolar(magnitude[n], getData()[n].getPhase());
   }
 }
