+++ LibSource/FloatMatrix.cpp
@@ -7,13 +7,13 @@ FloatMatrix::FloatMatrix(){
   instance.numCols = 0;
   instance.pData = NULL;
 }
-FloatMatrix::FloatMatrix(float* dt, int rs, int cs){
+FloatMatrix::FloatMatrix(float* dt, size_t rs, size_t cs){
   arm_mat_init_f32(&instance, rs, cs, dt);
 }
 #else
 FloatMatrix::FloatMatrix()
   : data(NULL), rows(0), columns(0) {}
-FloatMatrix::FloatMatrix(float* dt, int rs, int cs)
+FloatMatrix::FloatMatrix(float* dt, size_t rs, size_t cs)
   : data(dt), rows(rs), columns(cs) {}
 #endif
 
@@ -21,7 +21,7 @@ void FloatMatrix::add(FloatMatrix operand2, FloatMatrix destination){
 #ifdef ARM_CORTEX
   arm_mat_add_f32(&instance, &operand2.instance, &destination.instance);
 #else
-#error todo
+#warning todo
 #endif
 }
   
@@ -30,9 +30,9 @@ void FloatMatrix::add(FloatMatrix operand2){
 }
   
 void FloatMatrix::add(float scalar){
-  int size = getSize();
+  size_t size = getSize();
   float* data = getData();
-  for(int i=0; i<size; i++)
+  for(size_t i=0; i<size; i++)
       data[i] += scalar;
 }
 
@@ -40,7 +40,7 @@ void FloatMatrix::multiply(FloatMatrix operand2, FloatMatrix destination){
 #ifdef ARM_CORTEX
   arm_mat_mult_f32(&instance, &operand2.instance, &destination.instance);
 #else
-#error todo
+#warning todo
 #endif
 }
   
@@ -49,20 +49,20 @@ void FloatMatrix::multiply(FloatMatrix operand2){
 }
   
 void FloatMatrix::multiply(float scalar){
-  int size = getSize();
+  size_t size = getSize();
   float* data = getData();
-  for(int i=0; i<size; i++)
+  for(size_t i=0; i<size; i++)
       data[i] *= scalar;
 }
 
 void FloatMatrix::setAll(float value){
-  int size = getSize();
+  size_t size = getSize();
   float* data = getData();
-  for(int i=0; i<size; i++)
+  for(size_t i=0; i<size; i++)
     data[i] = value;
 }
 
- FloatMatrix FloatMatrix::create(int m, int n){
+ FloatMatrix FloatMatrix::create(size_t m, size_t n){
   return FloatMatrix(new float[m*n], m, n);
 }
